wrapped up first pass on segment provider

+++ server/segment_provider/sim/api_model.go
@@ -2,7 +2,7 @@ package sim
 
 type vehicleHist struct {
   //how long ago in timesteps the vehicle was on this segment
-  Elapsed int `json:"elapsed"`
+  Elapsed float64 `json:"elapsed"`
   //the segment id
   SegmentId string `json:"id"`
 }
